Creating new invoice now correctly sets largest number

+++ shared/utils/firebase.ts
@@ -91,21 +91,23 @@ export const createInvoice = (invoice: Invoice) => {
 };
 
 export const deleteInvoice = async (invoice_no: string) => {
-	const invoiceQuery = firestore
-		.collection("invoices")
-		.where("invoice_no", "==", invoice_no)
-		.where("owner", "==", getCurrentUser()?.uid);
+	if (confirm("Are you sure you want to delete this invoice?")) {
+		const invoiceQuery = firestore
+			.collection("invoices")
+			.where("invoice_no", "==", invoice_no)
+			.where("owner", "==", getCurrentUser()?.uid);
 
-	await invoiceQuery.get().then((querySnapshot) => {
-		querySnapshot.forEach((doc) => {
-			doc.ref
-				.delete()
-				.then(() => toast.error("Invoice deleted"))
-				.catch((error) => {
-					console.error("Error removing document: ", error);
-				});
+		await invoiceQuery.get().then((querySnapshot) => {
+			querySnapshot.forEach((doc) => {
+				doc.ref
+					.delete()
+					.then(() => toast.error("Invoice deleted"))
+					.catch((error) => {
+						console.error("Error removing document: ", error);
+					});
+			});
 		});
-	});
+	}
 };
 
 export const getLastInvoiceDetails = async () => {
@@ -124,6 +126,33 @@ export const getLastInvoiceDetails = async () => {
 		});
 
 	return lastInvoice;
+};
+
+export const getHighestInvoiceNumber = async () => {
+	const invoices: Invoice[] = [];
+
+	await firestore
+		.collection("invoices")
+		.where("owner", "==", auth.currentUser?.uid)
+		.orderBy("invoice_no", "desc")
+		.get()
+		.then((querySnapshot) => {
+			querySnapshot.forEach((doc: firebase.firestore.DocumentData) => {
+				const invoice: Invoice = doc.data();
+
+				invoices.push(invoice);
+			});
+		});
+
+	let invoiceNumbers = invoices.map((invoice) => invoice.invoice_no);
+	invoiceNumbers = invoiceNumbers.sort((a, b) =>
+		parseInt(a.replace(/([A-Za-z])+/g, ""), 10) <
+		parseInt(b.replace(/([A-Za-z])+/g, ""), 10)
+			? 1
+			: 0
+	);
+
+	return invoiceNumbers[0];
 };
 
 //
